Operation not_found result block is implemented
When running an operation it is possible to define result.not_found { ... }
to handle operations with :not_found terminus.

If the operation support the terminus and the result block is not
implemented it will raise ActiveRecord::RecordNotFound exception, and
if it is implemented the block will be executed with the custom logic
  provided.

If the operation doesn't not support it, nothing will happen.